[file.c] Unified similar functions file related
