+++ src/core/parser/elements/textures.c
@@ -6,7 +6,7 @@
 /*   By: ana-cast <ana-cast@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 17:16:11 by ana-cast          #+#    #+#             */
-/*   Updated: 2025/06/10 17:31:25 by ana-cast         ###   ########.fr       */
+/*   Updated: 2025/06/11 18:43:49 by ana-cast         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,43 +70,6 @@ bool	add_texture(mlx_texture_t *texture, t_hlist **txt_lst)
 	return (0);
 }
 
-int	number_of_occurrences(char *str, int c)
-{
-	int	i;
-	int	n;
-
-	i = -1;
-	n = 0;
-	while ((char)str[++i])
-	{
-		if (str[i] == (char)c)
-			n++;
-	}
-	if ((char)c == '\0')
-		return (1);
-	return (n);
-}
-
-bool	check_extension(char *path, char *extension)
-{
-	char	*ext;
-	char	*file;
-
-	if (!path || ft_strlen(path) < ft_strlen(extension))
-		return (1);
-	file = ft_strrchr(path, '/');
-	if (!file)
-		file = path;
-	if (number_of_occurrences(file, '.') != 1)
-		return (1);
-	ext = ft_strrchr(file, '.');
-	if (!ext)
-		return (1);
-	if (ft_strncmp(ext, extension, ft_strlen(extension) + 1) != 0)
-		return (1);
-	return (0);
-}
-
 char	*path_from_texture(t_game *game, t_parser *parser)
 {
 	char	*path;
@@ -150,33 +113,3 @@ void	parse_texture_line(t_game *game, t_direction dir)
 	parser->textures[dir] = true;
 	free_array(&parser->element);
 }
-
-/*
-void	parse_texture_line(t_game *game, char *line, t_direction dir)
-{
-	char	**content;
-	char	*trimmed;
-	int		fd;
-
-	if (dir == INVALID)
-		error_exit(game, E_TEX_INVALID, line);
-	//else if (ALLOW_SPRITES && game->parser_state->textures[dir] == true) // check allow sprites
-	//	error_exit(game, E_TEX_DUP, get_direction_name(dir));
-	trimmed = ft_strtrim(line, "\n");
-	if (!trimmed)
-		error_exit(game, E_MEM_ALLOC, "trimming texture");
-	content = ft_split(trimmed, ' ');
-	if (!free_str(&trimmed) && !content)
-		error_exit(game, E_MEM_ALLOC, "parsing texture");
-	if (!content[1])
-		error_exit(game, E_TEX_MISSING, content[0]);
-	fd = open(content[1], O_RDONLY);
-	if (fd < 0)
-		error_exit(game, E_TEX_LOAD, content[1]);
-	close(fd);
-	if (add_texture(mlx_load_png(content[1]), &game->graphics->textures_lst[dir]))
-		error_exit(game, E_TEX_LOAD, content[1]);
-	game->parser_state->textures[dir] = true;
-	free_array(&content);
-}
-*/
